LMSA-7244 Make sure userInit isn't null before trying to find a shortcut

+++ src/main/java/edu/iu/uits/lms/gct/services/GoogleCourseToolsService.java
@@ -1495,9 +1495,11 @@ public class GoogleCourseToolsService implements InitializingBean {
       for (String userEmail : toRemoveFromAll) {
          removeMemberFromGroup(courseDetail.getAllGroupEmail(), userEmail);
          UserInit userInit = userInitMap.computeIfAbsent(userEmail, key -> userInitRepository.findByGoogleLoginIdAndEnv(key, toolConfig.getEnv()));
-         File shortcut = findShortcutForTarget(courseFolder.getName(), courseFolder.getId(), userInit.getFolderId(), null);
-         if (shortcut != null) {
-            deleteFolder(shortcut.getId());
+         if (userInit != null) {
+            File shortcut = findShortcutForTarget(courseFolder.getName(), courseFolder.getId(), userInit.getFolderId(), null);
+            if (shortcut != null) {
+               deleteFolder(shortcut.getId());
+            }
          }
 
          //remove dropbox perms if it's a student
